Extract specific methods of array and a hash
Since we have some methods on formatter that is used only on the array
and hash moved to a module.
