+++ lib/awesome_print/formatters/hash.rb
@@ -1,6 +1,9 @@
+require 'awesome_print/formatters/enumerable'
+
 module AwesomePrint
   module Formatters
     class Hash < Base
+      include Enumerable
 
       def call
         return empty_format if object.empty?
@@ -44,7 +47,7 @@ module AwesomePrint
             end
           end
 
-          @data = formatter.limited(@data, width, :hash => true) if formatter.should_be_limited?
+          @data = limited(@data, width, :hash => true) if should_be_limited?
         end
 
         def multiline_format
